+++ packages/orq-rc/src/models/operations/updatebudget.ts
@@ -436,7 +436,7 @@ export const UpdateBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-21T07:50:16.259Z",
+    "2025-11-21T10:28:39.070Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -473,7 +473,7 @@ export const UpdateBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => UpdateBudgetConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-21T07:50:16.259Z"))
+  updated: z.date().default(() => new Date("2025-11-21T10:28:39.070Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {